1. remove images 2. create admin login 3. improvement view image api to return fallback image

+++ src/Migrations/2020_11_18_015852_create_user_roles.php
@@ -14,7 +14,8 @@ class CreateUserRoles extends Migration
     public function up()
     {
         Schema::create('user_roles', function (Blueprint $table) {
-            $table->unsignedInteger('user_id')->index();
+            // $table->unsignedInteger('user_id')->index();
+            $table->unsignedBigInteger('user_id')->index();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->unsignedInteger('role_id')->index();
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
